gas_utility_project/requirements.txt
boto
asgiref==3.6.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==23.2
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Naked==0.1.32
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
soupsieve==2.4.1
sqlparse==0.4.4
status==0.2.5
tzdata==2023.3
urllib3==2.0.2
django-storages
drf-spectacular